fix: Allow host and port to be specified
If I define a env var `REDIS_PORT=1234` and put a breakpoint on line 37 I can see `process.env.REDIS_PORT` equals `1234` but `DEFAULT_PORT` is `6397` so the port is not actually set.

This change allows it to be set via optional constructor args. 

I'm not sure anyone can specify a non-default port at the moment, but just in case, my change doesn't affect existing behaviour.